+++ extensions/visual-studio-code-recent-projects/src/types.ts
@@ -1,44 +1,49 @@
+import { existsSync } from "fs";
+import { URL } from "url";
+
 export interface FileEntry {
-  id: string;
   fileUri: string;
 }
 
 export function isFileEntry(entry: EntryLike): entry is FileEntry {
-  const { id } = entry as FileEntry;
-  return id === "openRecentFile";
+  const { fileUri } = entry as FileEntry;
+  return fileUri !== undefined && existsSync(new URL(fileUri)) && fileUri.indexOf(".code-workspace") === -1;
 }
 
 export interface FolderEntry {
-  id: string;
   folderUri: string;
-  scheme: string;
 }
 
 export function isFolderEntry(entry: EntryLike): entry is FolderEntry {
-  const { id, scheme } = entry as FolderEntry;
-  return id === "openRecentFolder" && scheme !== "vscode-remote";
+  const { folderUri } = entry as FolderEntry;
+  return folderUri !== undefined && existsSync(new URL(folderUri));
 }
 
 export interface WorkspaceEntry {
-  id: string;
-  fileUri: string;
+  workspace: {
+    configPath: string;
+  };
 }
 
 export function isWorkspaceEntry(entry: EntryLike): entry is WorkspaceEntry {
-  const { id } = entry as WorkspaceEntry;
-  return id === "openRecentWorkspace";
+  const { workspace } = entry as WorkspaceEntry;
+
+  return (
+    workspace !== undefined &&
+    existsSync(new URL(workspace.configPath)) &&
+    workspace.configPath.indexOf(".code-workspace") !== -1
+  );
 }
 
 export interface RemoteEntry {
-  id: string;
   folderUri: string;
-  scheme: string;
+  remoteAuthority: string;
   label: string;
 }
 
 export function isRemoteEntry(entry: EntryLike): entry is RemoteEntry {
-  const { id, scheme } = entry as RemoteEntry;
-  return id === "openRecentFolder" && scheme === "vscode-remote";
+  const { folderUri, remoteAuthority } = entry as RemoteEntry;
+  return folderUri !== undefined && remoteAuthority !== undefined;
 }
 
 export type EntryLike = FolderEntry | FileEntry | WorkspaceEntry | RemoteEntry;
@@ -47,33 +52,7 @@ export enum VSCodeBuild {
   Code = "Code",
   Insiders = "Code - Insiders",
 }
+
 export interface Preferences {
   build: VSCodeBuild;
-}
-
-export const RecentOpenedId = "submenuitem.35";
-export enum RecentOpenedItemId {
-  Folder = "openRecentFolder",
-  File = "openRecentFile",
-  Workspace = "openRecentWorkspace",
-  Other = "useless",
-}
-
-interface RecentOpenItem {
-  id: RecentOpenedItemId.Folder | RecentOpenedItemId.File | RecentOpenedItemId.Workspace;
-  uri: {
-    path: string;
-    scheme: string;
-    external: string;
-  };
-  enabled: boolean;
-  label: string;
-}
-
-export interface lastKnownMenubarItems {
-  id: string;
-  label: string;
-  submenu?: {
-    items: Array<RecentOpenItem>;
-  };
 }