+++ extensions/visual-studio-code-recent-projects/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Visual Studio Code Search Recent Projects Changelog
 
+## [Fixes] - 2022-05-25
+
+- Reverted back to using the VS Code SQLite DB to load recent projects.
+- Fixed an issue where only 10 recent projects would load.
+
 ## [Updates] - 2022-05-25
 
 - Fixed a bug where it was unable to open project that has space in path